fix: height changes not evaluated (#132)
I'm having intermittent issues with the bottom sheet not appearing.

I tracked the issue down to the `height` field. The onLayout callback (`handleFullHeader`) is triggered with the correct dimensions and the `height` AnimatedValue is updated, but the nodes that use this value (opacity/translateY) are never re-evaluated unless something else manually triggers a re-render.

This diff seems to fix the problem, although I'm not totally sure why.

I did also try replacing the callback with `Animated.event()` but this does not seem to work (https://github.com/software-mansion/react-native-reanimated/issues/79).